Public-portfolio/Energy/power_price_forecasting - In_dev/power_price_forecasting_v2/requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
lightgbm>=4.0
matplotlib>=3.7
seaborn>=0.12
pyyaml>=6.0
pyarrow>=14.0
holidays>=0.40
requests>=2.31
yfinance>=0.2.40
jupyter>=1.0
shap>=0.41
